backend/requirements.txt
# backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
httpx==0.25.2
websockets==11.0.3
uuid==1.30
openai==0.27.8
elevenlabs